requirements.txt
flask==3.1.0
psycopg2-binary==2.9.7
websockets==11.0.3
psutil==5.9.5
msgspec==0.18.4
pyzmq==25.1.1
zstandard==0.21.0
lz4==4.3.2
gunicorn==21.2.0
tqdm==4.66.1
ijson==3.3.0
flask-sock==0.7.0
Flask-SQLAlchemy==3.1.1
colorama==0.4.6
requests==2.31.0
discord.py>=2.0.0
table2ascii>=0.1.5
prettytable>=2.5.0
python-dotenv>=1.0.0
aiohttp>=3.8.0